Cluster E4: default_sc_group_size typed f32; ensureBuf parameter name parity
* `constants.default_sc_group_size` retyped `u8 = 4` → `f32 = 4.0` to
  match the v2 wire format. The frame writer field default and the
  parse-roundtrip test (`frame_format.zig`) both used `@floatFromInt`
  / `@as(u8, ...)` wrappers that were noise; collapsed to the bare
  constant.
* `encode_context.ensureBuf(ptr, cur, needed)` parameter renamed
  `cur` → `current_size`, matching the decode-side
  `ensureDeviceBuf` parameter shape. Updated docstring to note the
  shape mismatch between the two surfaces (encode returns `bool`,
  decode returns `GpuError!void`).

Bench unchanged; tests green.

+++ src/format/streamlz_constants.zig
@@ -30,10 +30,11 @@ pub const safe_space: usize = 64;
 // ── Frame-header default ───────────────────────────────────────────────
 
 /// Default value the frame writer stamps into the `sc_group_size` slot
-/// when the encoder did not pick a non-default value. The current GPU
-/// encoder always overrides this via `resolveScGroupSize`; the default
-/// is kept for the canonical-header smoke tests.
-pub const default_sc_group_size: u8 = 4;
+/// when the encoder did not pick a non-default value. Typed as `f32` to
+/// match the v2 wire format; the current GPU encoder always overrides
+/// this via `resolveScGroupSize`, so the constant only flows through the
+/// canonical-header smoke tests.
+pub const default_sc_group_size: f32 = 4.0;
 
 // ── Match-window limits ────────────────────────────────────────────────
